move cw rum to App.tsx
