+++ src/chainlit/frontend/src/App.tsx
@@ -1,3 +1,4 @@
+import { AwsRum, AwsRumConfig } from 'aws-rum-web';
 import { useEffect } from 'react';
 import { Toaster } from 'react-hot-toast';
 import { RouterProvider } from 'react-router-dom';
@@ -5,8 +6,7 @@ import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 import { router } from 'router';
 import makeTheme from 'theme';
 
-import { Box, GlobalStyles } from '@mui/material';
-import { ThemeProvider } from '@mui/material';
+import { Box, GlobalStyles, ThemeProvider } from '@mui/material';
 
 import Hotkeys from 'components/Hotkeys';
 import SettingsModal from 'components/molecules/settingsModal';
@@ -19,6 +19,32 @@ import { settingsState } from 'state/settings';
 import { accessTokenState, roleState } from 'state/user';
 
 import './App.css';
+
+try {
+  const config: AwsRumConfig = {
+    sessionSampleRate: 1,
+    guestRoleArn:
+      'arn:aws:iam::852971690794:role/RUM-Monitor-us-east-1-852971690794-7248872325961-Unauth',
+    identityPoolId: 'us-east-1:415ccbd0-7a90-4a91-9d5c-86e777905441',
+    endpoint: 'https://dataplane.rum.us-east-1.amazonaws.com',
+    telemetries: ['performance', 'errors', 'http'],
+    allowCookies: true,
+    enableXRay: false
+  };
+
+  const APPLICATION_ID = '3bc175d7-be7d-46bd-ac23-b7553ed0f7f0';
+  const APPLICATION_VERSION = '1.0.0';
+  const APPLICATION_REGION = 'us-east-1';
+
+  const awsRum: AwsRum = new AwsRum(
+    APPLICATION_ID,
+    APPLICATION_VERSION,
+    APPLICATION_REGION,
+    config
+  );
+} catch (error) {
+  // Ignore errors thrown during CloudWatch RUM web client initialization
+}
 
 function App() {
   const client = useRecoilValue(clientState);
